requirements/base.txt
fastapi==0.103.1
SQLAlchemy==2.0.20
uvicorn==0.23.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pydantic[email]
python-jose==3.3.0